+++ src/CustomizeBuildings/Properties/AssemblyInfo.cs
@@ -2,5 +2,5 @@ using System.Reflection;
 
 [assembly: AssemblyTitle("CustomizeBuildings")]
 [assembly: AssemblyProduct("CustomizeBuildings")]
-[assembly: AssemblyVersion("1.0.77.2")]
-[assembly: AssemblyFileVersion("1.0.77.2")]
+[assembly: AssemblyVersion("1.0.78.1")]
+[assembly: AssemblyFileVersion("1.0.78.1")]